NAPRAWA ZAKOKRAGLANIE IMPOORT XLS BANKI

+++ npkpir_23/src/java/xls/X.java
@@ -93,7 +93,7 @@ public class X {
                     zwrot = Double.valueOf(pobrane);
                     break;
                 default:
-                    zwrot = (int) cell.getNumericCellValue();
+                    zwrot = cell.getNumericCellValue();
                     break;
             }
         }
